Update ui.R
Changing the column name of the data set to make it easier to code. I did this because every time I tried to save it the "i" would change to a question mark so it wouldn't save correctly. Changing the training and holdout data. Also adding a multiple models to test which one best predicts best. Finally testing the model using RMSE to see which one predicts closest to the training data . The model didn't work when I tried the search and neural so that is why they are commented out.